+++ utils/types/jec-glasscat-config.d.ts
@@ -5,11 +5,9 @@
  * This is a part of the JEC projects: <http://jecproject.org>
  */
 
-
-
 declare module "jec-glasscat-config" {
 
-    import { LogLevelString, CacheControlPolicy } from "jec-commons";
-    import { AuthMethod, RealmType, SessionStorageType } from "jec-exchange";
+import { LogLevelString, CacheControlPolicy } from "jec-commons";
+import { AuthMethod, RealmType, SessionStorageType } from "jec-exchange";
     
 /*PLACEHOLDER*/}