+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 # JEC GlassCat Config Project: Update Release Notes
 
+<a name="jec-glasscat-config-1.0.2"></a>
+## **1.0.2** (2018-07-15)
+
+### Bug Fixes
+
+- Moving main js file
+
 <a name="jec-glasscat-config-1.0.1"></a>
 ## **1.0.1** (2018-07-15)
 
 ### Bug Fixes
 
 - Fixing definition types imports
-- Fixing main js filereference
+- Fixing main js file reference
 
 ### Features